+++ packages/moonstone/CHANGELOG.md
@@ -7,7 +7,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 ### Added
 
 - `moonstone/Panels` slot `<controls>` to easily add custom controls next to the Panels' "close" button
-- `moonstone/Tooltip` and `moonstone/TooltipDecorator` support for relative positioning via the `tooltipRelative` prop. This is an advanced feature and requires a container with specific rules. See documentation for details.
+- `moonstone/TooltipDecorator` prop `tooltipRelative` and `moonstone/TooltipDecorator.Tooltip` prop `relative`  to support for relative positioning.
 
 ### Removed